% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendsort-package.R
\docType{data}
\name{sample_tcga}
\alias{sample_tcga}
\title{Sample data matrix from the integrated pathway analysis of gastric cancer from the Cancer Genome Atlas (TCGA) study}
\format{A data frame with 215 rows and 117 variables}
\usage{
data(sample_tcga)
}
\description{
a multivariate table obtained from the integrated pathway analysis of gastric cancer from the Cancer Genome Atlas (TCGA) study.
In this data set, each column represents a pathway consisting of a set of genes and each row represents a cohort of samples based
on specific clinical or genetic features. For each pair of a pathway and a feature, a continuous value of between 1 and -1 is
assigned to score positive or negative association, respectively.
}
\details{
We would like to thank Sheila Reynolds and Vesteinn Thorsson from the Institute for Systems Biology for sharing this sample data set.
}
\keyword{datasets}

