% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendsort-package.R
\docType{package}
\name{dendsort-package}
\alias{dendsort-package}
\title{Modular Leaf Ordering Methods for Dendrogram Nodes}
\description{
Modular Leaf Ordering Methods for Dendrogram Nodes
}
\details{
This package includes functions to optimize ordering of nodes in a dendrogram, without affecting the meaning of the dendrogram. 
A dendrogram can be sorted based on the average distance of subtrees, or based on the smallest distance value. These sorting 
methods improve readability and interpretability of tree structure, especially for tasks such as comparison of different 
distance measures or linkage types and identification of tight clusters and outliers. As a result, it also introduces more 
meaningful reordering for a coupled heatmap visualization.
}
\author{
Ryo Sakai \email{ryo.sakai@esat.kuleuven.be}
}
\keyword{package}

